backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2

# PDF Generation
reportlab==4.0.7
pdfkit==1.0.0

# Database
psycopg2-binary==2.9.9

# Async tasks (optional for now)
# celery==5.3.4
# redis==5.0.1

# Development
django-extensions==3.2.3
